+++ projects/framework-maya/source/ftrack_framework_maya/__init__.py
@@ -9,8 +9,6 @@ import platform
 import maya.cmds as cmds
 import maya.mel as mm
 
-import ftrack_api
-
 from ftrack_framework_core.host import Host
 from ftrack_framework_core.event import EventManager
 from ftrack_framework_core.client import Client
@@ -24,6 +22,8 @@ from ftrack_utils.extensions.environment import (
 )
 from ftrack_utils.usage import set_usage_tracker, UsageTracker
 from ftrack_utils.actions import register_remote_action
+
+from ftrack_utils.session import create_api_session
 
 from ftrack_framework_maya.utils import (
     dock_maya_right,
@@ -145,7 +145,7 @@ def bootstrap_integration(framework_extensions_path):
         f' {framework_extensions_path}'
     )
     # Create ftrack session and instantiate event manager
-    session = ftrack_api.Session(auto_connect_event_hub=True)
+    session = create_api_session(auto_connect_event_hub=True)
     event_manager = EventManager(
         session=session, mode=constants.event.LOCAL_EVENT_MODE
     )